+++ src/lib/_test_utils.ts
@@ -3,9 +3,10 @@ import bufferToArray from 'buffer-to-arraybuffer';
 import { createHash } from 'crypto';
 import * as pkijs from 'pkijs';
 
-import { generateRSAKeyPair, getPublicKeyDigestHex } from './crypto/keys/generation';
+import { generateRSAKeyPair } from './crypto/keys/generation';
 import { Certificate } from './crypto/x509/Certificate';
 import FullCertificateIssuanceOptions from './crypto/x509/FullCertificateIssuanceOptions';
+import { getPublicKeyDigestHex } from './crypto/keys/digest';
 
 export const CRYPTO_OIDS = {
   AES_CBC_128: '2.16.840.1.101.3.4.1.2',
