+++ src/lib/crypto/keys/generation.spec.ts
@@ -1,15 +1,5 @@
-import { createHash } from 'crypto';
-
-import { arrayBufferFrom, sha256Hex } from '../../_test_utils';
 import { ECDHCurveName, HashingAlgorithm, RSAModulus } from '../algorithms';
-import {
-  generateECDHKeyPair,
-  generateRSAKeyPair,
-  getIdFromIdentityKey,
-  getPublicKeyDigest,
-  getPublicKeyDigestHex,
-  getRSAPublicKeyFromPrivate,
-} from './generation';
+import { generateECDHKeyPair, generateRSAKeyPair, getRSAPublicKeyFromPrivate } from './generation';
 import { RsaPssPrivateKey } from './PrivateKey';
 import { MockRsaPssProvider } from '../webcrypto/_test_utils';
 import { derSerializePublicKey } from './serialisation';
@@ -164,59 +154,5 @@ describe('getRSAPublicKeyFromPrivate', () => {
     const publicKey = await getRSAPublicKeyFromPrivate(keyPair.privateKey);
 
     expect(publicKey.usages).toEqual(['verify']);
-  });
-});
-
-describe('getPublicKeyDigest', () => {
-  test('SHA-256 digest should be returned in hex', async () => {
-    const keyPair = await generateRSAKeyPair();
-
-    const digest = await getPublicKeyDigest(keyPair.publicKey);
-
-    expect(Buffer.from(digest)).toEqual(
-      createHash('sha256')
-        .update(await derSerializePublicKey(keyPair.publicKey))
-        .digest(),
-    );
-  });
-
-  test('Public key should be extracted first if input is private key', async () => {
-    const mockPublicKeySerialized = arrayBufferFrom('the public key');
-    const provider = new MockRsaPssProvider();
-    provider.onExportKey.mockResolvedValue(mockPublicKeySerialized);
-    const privateKey = new RsaPssPrivateKey('SHA-256', provider);
-
-    const digest = await getPublicKeyDigest(privateKey);
-
-    expect(Buffer.from(digest)).toEqual(
-      createHash('sha256').update(Buffer.from(mockPublicKeySerialized)).digest(),
-    );
-  });
-});
-
-test('getPublicKeyDigestHex should return the SHA-256 hex digest of the public key', async () => {
-  const keyPair = await generateRSAKeyPair();
-
-  const digestHex = await getPublicKeyDigestHex(keyPair.publicKey);
-
-  expect(digestHex).toEqual(sha256Hex(await derSerializePublicKey(keyPair.publicKey)));
-});
-
-describe('getIdFromIdentityKey', () => {
-  test('Id should be computed from identity key', async () => {
-    const keyPair = await generateRSAKeyPair();
-
-    const id = await getIdFromIdentityKey(keyPair.publicKey);
-
-    expect(id).toEqual('0' + sha256Hex(await derSerializePublicKey(keyPair.publicKey)));
-  });
-
-  test('DH keys should be refused', async () => {
-    const keyPair = await generateECDHKeyPair();
-
-    await expect(getIdFromIdentityKey(keyPair.publicKey)).rejects.toThrowWithMessage(
-      Error,
-      'Only RSA keys are supported (got ECDH)',
-    );
   });
 });
