+++ src/lib/nodes/Gateway.spec.ts
@@ -1,13 +1,14 @@
 import { addDays, setMilliseconds } from 'date-fns';
 
 import { reSerializeCertificate } from '../_test_utils';
-import { generateRSAKeyPair, getIdFromIdentityKey } from '../crypto/keys/generation';
+import { generateRSAKeyPair } from '../crypto/keys/generation';
 import { Certificate } from '../crypto/x509/Certificate';
 import { MockKeyStoreSet } from '../keyStores/testMocks';
 import { CertificationPath } from '../pki/CertificationPath';
 import { issueGatewayCertificate } from '../pki/issuance';
 import { Gateway } from './Gateway';
 import { StubVerifier } from './signatures/_test_utils';
+import { getIdFromIdentityKey } from '../crypto/keys/digest';
 
 let nodeId: string;
 let nodePrivateKey: CryptoKey;
