+++ src/lib/nodes/Node.spec.ts
@@ -2,7 +2,7 @@ import { addDays, setMilliseconds, subDays } from 'date-fns';
 
 import { arrayBufferFrom, expectArrayBuffersToEqual, reSerializeCertificate } from '../_test_utils';
 import { SessionEnvelopedData } from '../crypto/cms/envelopedData';
-import { generateRSAKeyPair, getIdFromIdentityKey } from '../crypto/keys/generation';
+import { generateRSAKeyPair } from '../crypto/keys/generation';
 import { Certificate } from '../crypto/x509/Certificate';
 import { MockKeyStoreSet } from '../keyStores/testMocks';
 import { ParcelDeliverySigner, ParcelDeliveryVerifier } from '../messages/bindings/signatures';
@@ -12,6 +12,7 @@ import { StubMessage } from '../ramf/_test_utils';
 import { StubNode } from './_test_utils';
 import { InvalidMessageError } from '../messages/InvalidMessageError';
 import { derSerializePublicKey } from '../crypto/keys/serialisation';
+import { getIdFromIdentityKey } from '../crypto/keys/digest';
 
 let nodeId: string;
 let nodePrivateKey: CryptoKey;
