+++ src/lib/nodes/channels/Channel.ts
@@ -1,11 +1,11 @@
 import { SessionEnvelopedData } from '../../crypto/cms/envelopedData';
-import { getIdFromIdentityKey } from '../../crypto/keys/generation';
 import { Certificate } from '../../crypto/x509/Certificate';
 import { KeyStoreSet } from '../../keyStores/KeyStoreSet';
 import { PayloadPlaintext } from '../../messages/payloads/PayloadPlaintext';
 import { Recipient } from '../../messages/Recipient';
 import { NodeError } from '../errors';
 import { NodeCryptoOptions } from '../NodeCryptoOptions';
+import { getIdFromIdentityKey } from '../../crypto/keys/digest';
 
 export abstract class Channel {
   // noinspection TypeScriptAbstractClassConstructorCanBeMadeProtected
