+++ src/lib/nodes/channels/GatewayChannel.spec.ts
@@ -9,7 +9,7 @@ import {
 } from '../../_test_utils';
 import { EnvelopedData, SessionEnvelopedData } from '../../crypto/cms/envelopedData';
 import { SignatureOptions } from '../../crypto/cms/SignatureOptions';
-import { generateRSAKeyPair, getIdFromIdentityKey } from '../../crypto/keys/generation';
+import { generateRSAKeyPair } from '../../crypto/keys/generation';
 import { Certificate } from '../../crypto/x509/Certificate';
 import { MockKeyStoreSet } from '../../keyStores/testMocks';
 import { Cargo } from '../../messages/Cargo';
@@ -22,6 +22,7 @@ import { SessionKey } from '../../SessionKey';
 import { SessionKeyPair } from '../../SessionKeyPair';
 import { NodeCryptoOptions } from '../NodeCryptoOptions';
 import { GatewayChannel } from './GatewayChannel';
+import { getIdFromIdentityKey } from '../../crypto/keys/digest';
 
 const MESSAGE = Buffer.from('This is a message to be included in a cargo');
