+++ src/lib/nodes/channels/PrivateGatewayChannel.spec.ts
@@ -1,10 +1,6 @@
 import { addDays, setMilliseconds, subMinutes, subSeconds } from 'date-fns';
 
-import {
-  generateRSAKeyPair,
-  getIdFromIdentityKey,
-  getRSAPublicKeyFromPrivate,
-} from '../../crypto/keys/generation';
+import { generateRSAKeyPair, getRSAPublicKeyFromPrivate } from '../../crypto/keys/generation';
 import { Certificate } from '../../crypto/x509/Certificate';
 import { MockKeyStoreSet } from '../../keyStores/testMocks';
 import { CertificationPath } from '../../pki/CertificationPath';
@@ -12,6 +8,7 @@ import { issueGatewayCertificate } from '../../pki/issuance';
 import { NodeCryptoOptions } from '../NodeCryptoOptions';
 import { PrivateGatewayChannel } from './PrivateGatewayChannel';
 import { derSerializePublicKey } from '../../crypto/keys/serialisation';
+import { getIdFromIdentityKey } from '../../crypto/keys/digest';
 
 let internetGatewayId: string;
 let internetGatewayPublicKey: CryptoKey;
